Various build script changes to simplify building everything

+++ htop/build.sh
@@ -3,8 +3,9 @@ set -eu
 
 ANDROID_API=23
 ARCH=arm64
-NDK_ROOT=${NDK_ROOT:-/opt/android-ndk}
+NDK_ROOT="${NDK_ROOT:-/opt/android-ndk}"
 
+export PROJECT=htop
 
 TOP="$(realpath "$(dirname "$0")")"
 cd "${TOP}"
@@ -35,7 +36,7 @@ export HOST="${GCC_ARCH}-linux-android$EABI"
 
 TOOLCHAIN="${NDK_ROOT}/toolchains/${HOST}-4.9/prebuilt/linux-x86_64"
 
-NCURSES_INSTALL="$TOP/../ncurses/install_dir/$ARCH/data/local/ncurses"
+NCURSES_INSTALL="$OUTDIR/../ncurses/$ARCH/$PREFIX"
 NCURSES_INCLUDE="$NCURSES_INSTALL/include"
 NCURSES_LIB="$NCURSES_INSTALL/lib"
 
@@ -46,25 +47,27 @@ export CFLAGS="--sysroot=${SYSROOT} -I$NCURSES_INCLUDE -L$NCURSES_LIB -fPIE"
 export LDFLAGS="-L$NCURSES_LIB -pie"
 export QUILT_PATCHES="$TOP/patches"
 
-[ -e "$NCURSES_INCLUDE/ncursesw/ncurses.h" ] || "$TOP/../ncurses/build.sh"
+[ -e "$NCURSES_INCLUDE/ncursesw/ncurses.h" ] || env -i "$TOP/../ncurses/build.sh"
 
-PREFIX="/data/local/htop"
-
-export DESTDIR="$TOP/install_dir/$ARCH/"
+export DESTDIR="$OUTDIR/$ARCH/"
 
 # Cleanup old output
-rm -rf "$DESTDIR"
+rm -rf "${DESTDIR:?}"
 mkdir -p "$DESTDIR"
 
 cd src
 
-quilt push -a
+quilt push -a || [ $? == 2 ]
+
+make distclean
 
 ./autogen.sh
 
-./configure --host=${HOST} \
-		--prefix=${PREFIX} \
-		--exec-prefix=${PREFIX} \
+./configure --host="${HOST}" \
+		--prefix="${PREFIX}" \
+		--exec-prefix="${PREFIX}" \
+		--bindir="${PREFIX}/$BINDIR" \
+		--sbindir="${PREFIX}/$BINDIR" \
 		--enable-unicode \
 		--enable-shared=yes \
 		--enable-static=no
@@ -75,4 +78,4 @@ rm -rf "$DESTDIR/$PREFIX/share"
 
 quilt pop -af
 
-printf "\n\nBuild complete! See install_dir/$ARCH/\n"
+printf "\n\nBuild complete! See OUTDIR/%s/\n" "$ARCH"
