+++ nano/build.sh
@@ -3,8 +3,9 @@ set -eu
 
 ANDROID_API=23
 ARCH=arm64
-NDK_ROOT=${NDK_ROOT:-/opt/android-ndk}
+NDK_ROOT="${NDK_ROOT:-/opt/android-ndk}"
 
+export PROJECT=nano
 
 TOP="$(realpath "$(dirname "$0")")"
 cd "${TOP}"
@@ -35,42 +36,43 @@ export HOST="${GCC_ARCH}-linux-android$EABI"
 
 TOOLCHAIN="${NDK_ROOT}/toolchains/${HOST}-4.9/prebuilt/linux-x86_64"
 
-NCURSES_INSTALL="$TOP/../ncurses/install_dir/$ARCH/data/local/ncurses"
+NCURSES_INSTALL="$OUTDIR/../ncurses/$ARCH/$PREFIX"
 NCURSES_INCLUDE="$NCURSES_INSTALL/include"
 NCURSES_LIB="$NCURSES_INSTALL/lib"
 
 export CROSS_COMPILE="${HOST}-"
 export PATH="$TOOLCHAIN/bin:$PATH"
 export SYSROOT="${NDK_ROOT}/platforms/android-${ANDROID_API}/arch-${ARCH}"
-export CFLAGS="--sysroot=${SYSROOT} -I$NCURSES_INCLUDE -L$NCURSES_LIB -fPIE"
+export CFLAGS="--sysroot=${SYSROOT} -I$NCURSES_INCLUDE -I${NDK_ROOT}/sysroot/usr/include -L$NCURSES_LIB -fPIE"
 export LDFLAGS="-L$NCURSES_LIB -pie"
 
-[ -e "$NCURSES_INCLUDE/ncursesw/ncurses.h" ] || "$TOP/../ncurses/build.sh"
+[ -e "$NCURSES_INCLUDE/ncursesw/ncurses.h" ] || env -i "$TOP/../ncurses/build.sh"
 
-PREFIX="/data/local/nano"
-
-export DESTDIR="$TOP/install_dir/$ARCH/"
+export DESTDIR="$OUTDIR/$ARCH/"
 
 # Cleanup old output
-rm -rf "$DESTDIR"
-mkdir -p install_dir/$ARCH
+rm -rf "${DESTDIR:?}"
+mkdir -p "$DESTDIR"
 
 cd src
 
-gnulib-tool --import glob unistr/m8-mblen
+gnulib-tool --import glob unistr/u8-mblen
 
-patch -Np1 -i ../android.patch
+patch -Np1 -i ../android.patch || true
+patch -Np1 -i ../android-home.patch || true
 
 ./autogen.sh
 
-./configure --host=${HOST} \
-		--prefix=${PREFIX} \
-		--exec-prefix=${PREFIX} \
+./configure --host="${HOST}" \
+		--prefix="${PREFIX}" \
+		--exec-prefix="${PREFIX}" \
+		--bindir="${PREFIX}/$BINDIR" \
+		--sbindir="${PREFIX}/$BINDIR" \
 		--enable-utf8 \
 		ac_cv_func_nl_langinfo=yes ac_cv_func_mblen=yes
 
-make -j8
+make -j1
 
 make install-strip
 
-printf "\n\nBuild complete! See install_dir/$ARCH/\n"
+printf "\n\nBuild complete! See OUTDIR/%s/\n" "$ARCH"
